requirements.txt
discord.py>=2.6.0,<3
python-dotenv>=1.0.0,<2
PyNaCl>=1.5.0,<2

suntime>=1.2.5
geocoder>=1.38.1
humanfriendly>=8.2
durations_nlp>=1.0.1
expiringdict>=1.2.1
structlog>=22.3.0
python-json-logger>=2.0.4
fluent.runtime>=0.4.0
uvloop>=0.17.0; platform_system!='Windows'
